postController ready, tags ready

+++ src/main/java/com/skillbox/AndrewBlog/api/response/CommentEntityResponse.java
@@ -2,23 +2,23 @@ package com.skillbox.AndrewBlog.api.response;
 
 public class CommentEntityResponse {
 
-    private long id;
+    private int id;
     private long timestamp;
     private String text;
     private IdNamePhotoResponse user;
 
-    public CommentEntityResponse(long id, long timestamp, String text, IdNamePhotoResponse user) {
+    public CommentEntityResponse(int id, long timestamp, String text, IdNamePhotoResponse user) {
         this.id = id;
         this.timestamp = timestamp;
         this.text = text;
         this.user = user;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
